virtual environment and window menu
Due to problems after updating python to the newest version (3.13.3), virtual environment is set up. Now it is using version (3.10.4) and is compatible to both SpaCy and PySide6.

For UI, mainwindow is named to "requirement conflict detection software" and settings menu is successfully added. Menu content is still non functional as of now.

Also commented an attempt to creating tabs.