Prototype Commit
This is the commit when prototype is shown to supervisor. Contains functioning tab changes, functional table with delete and add row. Dummy warning that can be activated with button. Also non working menu.

Also included a testspacy.py that contains some codes with spacy used to demonstrate tokenization and similarity index.